@@ -1,6 +1,6 @@
 module github.com/gloflow/gloflow
 
-go 1.17
+go 1.19
 
 require (
 	github.com/PuerkitoBio/goquery v1.6.1
